+++ CHANGELOG.md
@@ -34,7 +34,7 @@
 
 ## 1.2.1
  - Renamed executable to "mutation_test" follow dart file conventions.
- - Added exclusion rules for dart import and export statements in the default ruleset.
+ - Added exclusion rules for dart import and export statements in the default rule set.
 
 ## 1.2.0